requirements.txt
absl-py==0.13.0
astunparse==1.6.3
beautifulsoup4==4.10.0
biopython==1.79
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
chex==0.0.7
contextlib2==21.6.0
cycler==0.11.0
dm-haiku==0.0.4
dm-tree==0.1.6
docker==5.0.0
filelock==3.4.2
flatbuffers==1.12
fonttools==4.28.5
gast==0.4.0
gdown==4.2.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.3
immutabledict==2.0.0
importlib-metadata==4.10.1
jax==0.2.14
jaxlib==0.1.74+cuda11.cudnn805
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
ml-collections==0.1.0
numpy==1.19.5
oauthlib==3.1.1
OpenMM==7.5.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
Pillow==9.0.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.0
six==1.15.0
soupsieve==2.3.1
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
toolz==0.11.2
torch==1.10.2
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.8
websocket-client==1.2.3
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.7.0
